pyproject.toml
[tool.poetry]
name = "llm-server"
version = "0.1.0"
description = ""
authors = ["nyaos <fjshsjf@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.90.0"
uvicorn = "^0.20.0"
gunicorn = "^20.1.0"
#langchain = "^0.0.100"
openai = "1.1.1"
websockets = "^10.4"
tiktoken = "0.3.3"
python-dotenv = "1.0.0"
numpy = "1.26.2"
websocket-client = "^1.7.0"


[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^1.0.0"
black = "^23.1.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"